swtich numRows field from int to long in druid adapter

+++ druid/src/main/java/org/apache/calcite/adapter/druid/DruidConnectionImpl.java
@@ -535,7 +535,7 @@ class DruidConnectionImpl implements DruidConnection {
     public List<String> intervals;
     public Map<String, JsonColumn> columns;
     public int size;
-    public int numRows;
+    public long numRows;
     public Map<String, JsonAggregator> aggregators;
   }
